% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{make_filename}
\alias{make_filename}
\title{Create FARS filename}
\usage{
make_filename(year, path = NULL)
}
\arguments{
\item{year}{An integer-valued, 4 number year (or a value that can be coerced to such)}

\item{path}{A character string representing the path that the FARS data is located in}
}
\value{
Returns a character string that represents a filename for
    accident data located in fars_data.zip
}
\description{
A helper function for creating a filename based on the desired accident year.
Useful in conjunction with `fars_read()`.
}
\examples{
make_filename(2013)

}

