% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read}
\alias{fars_read}
\title{Read FARS data}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{A character string for the name of a file in the `data` directory within fars_data.zip}
}
\value{
Returns a dataframe of accident data for a given filename.
    An error is produced if the file can't be located with the given `filename`.
}
\description{
Reads a file from `/data` inside fars_data.zip based on its filename.
Useful in conjunction with `make_filename()`.
}
\details{
Needs `dplyr` and `readr` packages available.
}

